scripts: add variable TERMUX_PKG_NO_REPLACE_GUESS_SCRIPTS

+++ scripts/build/termux_step_replace_guess_scripts.sh
@@ -1,5 +1,6 @@
 termux_step_replace_guess_scripts() {
 	[ "$TERMUX_PKG_METAPACKAGE" = "true" ] && return
+	[ "$TERMUX_PKG_NO_REPLACE_GUESS_SCRIPTS" = "true" ] && return
 
 	cd "$TERMUX_PKG_SRCDIR"
 	find . -name config.sub -exec chmod u+w '{}' \; -exec cp "$TERMUX_SCRIPTDIR/scripts/config.sub" '{}' \;
